requirements.txt
django-bootstrap4
django-registration
django-crispy-forms
